# server/openapi.yaml
openapi: 3.0.3
info:
  title: Metro Verify API
  description: >
    Digital Legal Metrology Verification & Lifecycle Management Platform (SIH26036).
    Covers Owner, Administrator, and LMO/GATC Officer endpoints, plus the AI chatbot.
    Endpoints marked "(stored procedure)" call a MySQL stored procedure for an atomic,
    multi-step operation rather than a plain Prisma query.
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development backend
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Application:
      type: object
      properties:
        applicationId: { type: string, example: "APP-1008" }
        instrumentId: { type: string }
        ownerId: { type: string }
        status:
          type: string
          enum: [SUBMITTED, PENDING, ASSIGNED, WITHDRAWN, COMPLETED]
        assignedOfficerId: { type: string, nullable: true }
...
    Error:
      type: object
      properties:
        error: { type: string }
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Owner login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: JWT issued }
        "401": { description: Invalid credentials, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/auth/officer/login:
    post:
      tags: [Auth]
      summary: LMO/GATC Officer login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                officerIdOrEmail: { type: string, example: "LMO-101" }
                password: { type: string }
      responses:
        "200": { description: JWT issued for the authenticated officer }
        "401": { description: Invalid credentials, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  # ---------------- OWNER: APPLICATIONS ----------------
  /api/applications:
    get:
      tags: [Owner - Applications]
      summary: List the logged-in owner's applications
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Application" }
    post:
      tags: [Owner - Applications]
      summary: Submit a single verification/re-verification request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                instrumentId: { type: string }
                verificationType: { type: string, enum: [VERIFICATION, RE_VERIFICATION] }
                preferredDate: { type: string, format: date }
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/applications/{id}:
    get:
      tags: [Owner - Applications]
      summary: Get a single application
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Application detail, content: { application/json: { schema: { $ref: "#/components/schemas/Application" } } } }
        "404": { description: Not found }
  /api/applications/{id}/withdraw:
    post:
      tags: [Owner - Applications]
      summary: Withdraw a submitted application (only if not yet completed)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: "Application status set to WITHDRAWN" }
        "409": { description: "Cannot withdraw — already completed/certified" }

  # ---------------- OWNER: BULK REQUESTS ----------------
...